Move migrations to AbstractPlugin
