+++ lib/core/AbstractPlugin.php
@@ -10,7 +10,7 @@ if( !class_exists( '\smartcat\core\AbstractPlugin' ) ) :
  * @package \smartcat\core
  * @author Eric Green <eric@smartcat.ca>
  */
-abstract class AbstractPlugin implements HookRegisterer, Plugin {
+abstract class AbstractPlugin implements HookRegisterer, HookSubscriber, Plugin {
     protected $url;
     protected $dir;
     protected $file;
@@ -26,6 +26,8 @@ abstract class AbstractPlugin implements HookRegisterer, Plugin {
         $this->id = $id;
         $this->file = $fs_context;
         $this->version = $version;
+
+        $this->add_api_subscriber( $this );
     }
 
     /**
@@ -109,6 +111,43 @@ abstract class AbstractPlugin implements HookRegisterer, Plugin {
         }
     }
 
+    public function perform_migrations() {
+        $current = get_option( $this->id . '_version', 0 );
+        $result = null;
+
+        $admin_notice = function ( $message, $class ) {
+            add_action( 'admin_notices', function () use ( $message, $class ) {
+                printf( '<div class="%1$s"><p>%2$s</p></div>', esc_attr( implode( ' ', $class ) ), $message );
+            } );
+        };
+
+        if( $this->version > $current ) {
+
+            // Perform migrations with the current version
+            foreach( glob($this->dir . 'migrations/migration-*.php' ) as $file ) {
+                $migration = include_once( $file );
+                $version = $migration->version();
+
+                if ( $version <= $this->version && $version > $current ) {
+                    $result = $migration->migrate();
+
+                    if ( !$result['success'] ) {
+                        $admin_notice( __( $result['message'], $this->id ), array( 'notice', 'notice-error', 'is-dismissible' ) );
+                        break;
+                    }
+                }
+            }
+
+            if( $result['success'] ) {
+                $admin_notice( __( $result['message'], $this->id ), array( 'notice', 'notice-success', 'is-dismissible' ) );
+
+                update_option( $this->id . '_version', $this->version );
+            }
+
+            error_log( $result['message'] );
+        }
+    }
+
     /**
      * The list of Components to instantiate.
      *
@@ -163,6 +202,10 @@ abstract class AbstractPlugin implements HookRegisterer, Plugin {
 
     public function version() {
         return $this->version;
+    }
+
+    public function subscribed_hooks( $hooks = array() )  {
+        return array_merge( array( 'init' => array( 'perform_migrations' ) ), $hooks );
     }
 
     /**
